+++ rust/src/lib.rs
@@ -160,16 +160,16 @@ mod ffi {
             radius: f64,
         );
 
-        fn generate(self: &SwervePathBuilderImpl, uuid: u32) -> Result<HolonomicTrajectory>;
-        fn enable_state_feedback(self: Pin<&mut SwervePathBuilderImpl>, callback: fn(HolonomicTrajectory, u32));
+        fn generate(self: &SwervePathBuilderImpl, uuid: i64) -> Result<HolonomicTrajectory>;
+        fn enable_state_feedback(self: Pin<&mut SwervePathBuilderImpl>, callback: fn(HolonomicTrajectory, i64));
 
         fn new_swerve_path_builder_impl() -> UniquePtr<SwervePathBuilderImpl>;
     }
 }
 
-static mut callback : Option<fn(HolonomicTrajectory, u32)->()> = None; 
+static mut callback : Option<fn(HolonomicTrajectory, i64)->()> = None; 
 
-pub fn set_progress_callback(cb: fn(HolonomicTrajectory, u32)->()) {
+pub fn set_progress_callback(cb: fn(HolonomicTrajectory, i64)->()) {
     unsafe {
     callback = Some(cb);
     }
@@ -380,7 +380,7 @@ impl SwervePathBuilder {
         );
     }
 
-    pub fn generate(&mut self, uuid: u32) -> Result<HolonomicTrajectory, String> {
+    pub fn generate(&mut self, uuid: i64) -> Result<HolonomicTrajectory, String> {
         unsafe {
         if callback.is_some() {
             let cb = callback.unwrap();
